refactor: extract date-time-picker logic into reusable mixin
