+++ packages/date-time-picker/test/typings/date-time-picker.types.ts
@@ -1,4 +1,12 @@
 import '../../vaadin-date-time-picker.js';
+import type { DisabledMixinClass } from '@vaadin/a11y-base/src/disabled-mixin.js';
+import type { FocusMixinClass } from '@vaadin/a11y-base/src/focus-mixin.js';
+import type { ElementMixinClass } from '@vaadin/component-base/src/element-mixin.js';
+import type { FieldMixinClass } from '@vaadin/field-base/src/field-mixin.js';
+import type { LabelMixinClass } from '@vaadin/field-base/src/label-mixin.js';
+import type { ValidateMixinClass } from '@vaadin/field-base/src/validate-mixin.js';
+import type { ThemableMixinClass } from '@vaadin/vaadin-themable-mixin/vaadin-themable-mixin.js';
+import type { DateTimePickerMixinClass } from '../../src/vaadin-date-time-picker-mixin.js';
 import type {
   DateTimePicker,
   DateTimePickerChangeEvent,
@@ -10,6 +18,16 @@ import type {
 const assertType = <TExpected>(actual: TExpected) => actual;
 
 const picker = document.createElement('vaadin-date-time-picker');
+
+// Mixins
+assertType<DateTimePickerMixinClass>(picker);
+assertType<DisabledMixinClass>(picker);
+assertType<ElementMixinClass>(picker);
+assertType<FieldMixinClass>(picker);
+assertType<FocusMixinClass>(picker);
+assertType<LabelMixinClass>(picker);
+assertType<ThemableMixinClass>(picker);
+assertType<ValidateMixinClass>(picker);
 
 picker.addEventListener('change', (event) => {
   assertType<DateTimePickerChangeEvent>(event);
